update logo and copyright; use the UXarray cross_section function; include GSI; update DAmonitor (#5)
* update build_book.sh and track logo.png
* fine tune build_book.sh; modify copyright through _config.yaml
* change slice to cross; update vslice_contour to use the latest UXarray cross_section() function
* update mpas_plotting.ipynb to match changes in the DAmonitor class
* update DAmonitor for GSI and import DAmonitor directly
* add gsi.ipynb and update other notebooks accordingly
* add gsi to docs and add .admin_update_book
* address linter comments

+++ docs/build_book.sh
@@ -5,10 +5,32 @@ doc_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 declare -A notebooks
 notebooks["mpas_plotting.ipynb"]="mpas_plotting.ipynb"
 notebooks["obs_exploring.ipynb"]="obs_exploring.ipynb"
+notebooks["gsi.ipynb"]="gsi.ipynb"
 
 
 ### users usually do not need to make changes below this line
 ### ========================================================================
+#
+# save the origin URL before running ghp-import, which has to use the origin remote
+save_origin=$(git remote get-url origin)
+#
+# get the remote url for current branch and set it as book_repo
+remote=$(git rev-parse --abbrev-ref --symbolic-full-name @{u} 2>/dev/null | cut -d'/' -f1)
+if [[ -z "${remote}" ]]; then
+  echo "Push current branch to a remote first and then run build_book.sh again"
+  exit 1
+fi
+book_repo=$(git remote get-url ${remote})
+#book_repo=git@github.com:pyDAmonitor/docs # use this line to overwrite the automatically-detected book_repo
+
+if [[ "${book_repo}" ==  *"NOAA-GSL/pyDAmonitor" ]] \
+   || [[ "${book_repo}" ==  *"pyDAmonitor/pyDAmonitor" ]]; then
+      echo "ERROR: current branch tracks the authoritative repository:"
+      echo "  ${book_repo}"
+      echo "users can ONLY push books to their own forks"
+      exit 1
+fi
+
 set -x
 ### always start from a clean notebook_docs/ directory
 rm -rf ${doc_dir}/notebook_docs
@@ -24,22 +46,6 @@ jupyter-book clean ${doc_dir}
 jupyter-book build ${doc_dir}
 
 ### push book to book_repo
-# save the origin URL and restore after ghp-import, which has to use the origin remote
-save_origin=$(git remote get-url origin)
-
-# get the remote url for current branch and set it as book_repo
-remote=$(git rev-parse --abbrev-ref --symbolic-full-name @{u} | cut -d'/' -f1)
-book_repo=$(git remote get-url ${remote})
-#book_repo= # use this line to overwrite the automatically-detected book_repo
-
-if [[ "${book_repo}" ==  *"NOAA-GSL/pyDAmonitor" ]] \
-   || [[ "${book_repo}" ==  *"pyDAmonitor/pyDAmonitor" ]]; then
-      echo "ERROR: current branch tracks the authoritative repository:"
-      echo "  ${book_repo}"
-      echo "users can ONLY push books to their own forks"
-      exit 1
-fi
-
 # change origin, ghp-import and then restore origin 
 git remote set-url origin ${book_repo}
 ghp-import -n -p -f ${doc_dir}/_build/html
